+++ packages/expect-utils/.npmignore
@@ -1,0 +1,7 @@
+**/__mocks__/**
+**/__tests__/**
+__typetests__
+src
+tsconfig.json
+tsconfig.tsbuildinfo
+api-extractor.json
