+++ packages/expect-utils/README.md
@@ -1,0 +1,5 @@
+# `@jest/expect-utils`
+
+This module exports some utils for the `expect` function used in [Jest](https://jestjs.io/).
+
+You probably don't want to use this package directly. E.g. if you're writing [custom matcher](https://jestjs.io/docs/expect#expectextendmatchers), you should use the injected [`this.equals`](https://jestjs.io/docs/expect#thisequalsa-b).
