+++ packages/expect-utils/src/index.ts
@@ -1,0 +1,13 @@
+/**
+ * Copyright (c) Facebook, Inc. and its affiliates. All Rights Reserved.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ *
+ */
+
+export {equals, isA} from './jasmineUtils';
+export type {EqualsFunction} from './jasmineUtils';
+export * from './utils';
+
+export type {Tester} from './types';
