+++ packages/expect/src/__tests__/extend.test.ts
@@ -6,11 +6,10 @@
  *
  */
 
+import {equals, iterableEquality, subsetEquality} from '@jest/expect-utils';
 import {alignedAnsiStyleSerializer} from '@jest/test-utils';
 import * as matcherUtils from 'jest-matcher-utils';
 import jestExpect from '../';
-import {equals} from '../jasmineUtils';
-import {iterableEquality, subsetEquality} from '../utils';
 
 expect.addSnapshotSerializer(alignedAnsiStyleSerializer);
