+++ packages/expect/src/matchers.ts
@@ -8,6 +8,17 @@
 
 /* eslint-disable local/ban-types-eventually */
 
+import {
+  arrayBufferEquality,
+  equals,
+  getObjectSubset,
+  getPath,
+  iterableEquality,
+  pathAsArray,
+  sparseArrayEquality,
+  subsetEquality,
+  typeEquality,
+} from '@jest/expect-utils';
 import {getType, isPrimitive} from 'jest-get-type';
 import {
   DIM_COLOR,
@@ -27,7 +38,6 @@ import {
   printWithType,
   stringify,
 } from 'jest-matcher-utils';
-import {equals} from './jasmineUtils';
 import {
   printCloseTo,
   printExpectedConstructorName,
@@ -39,16 +49,6 @@ import {
   printReceivedStringContainExpectedSubstring,
 } from './print';
 import type {MatchersObject} from './types';
-import {
-  arrayBufferEquality,
-  getObjectSubset,
-  getPath,
-  iterableEquality,
-  pathAsArray,
-  sparseArrayEquality,
-  subsetEquality,
-  typeEquality,
-} from './utils';
 
 // Omit colon and one or more spaces, so can call getLabelPrinter.
 const EXPECTED_LABEL = 'Expected';
