+++ packages/expect/src/toThrowMatchers.ts
@@ -8,6 +8,7 @@
 
 /* eslint-disable local/ban-types-eventually */
 
+import {isError} from '@jest/expect-utils';
 import {
   EXPECTED_COLOR,
   MatcherHintOptions,
@@ -35,7 +36,6 @@ import type {
   RawMatcherFn,
   SyncExpectationResult,
 } from './types';
-import {isError} from './utils';
 
 const DID_NOT_THROW = 'Received function did not throw';
