+++ packages/jest-snapshot/src/printSnapshot.ts
@@ -8,11 +8,7 @@
 /* eslint-disable local/ban-types-eventually */
 
 import chalk = require('chalk');
-// Temporary hack because getObjectSubset has known limitations,
-// is not in the public interface of the expect package,
-// and the long-term goal is to use a non-serialization diff.
-// Make sure to remove file from `exports` in `expect/package.json`.
-import {getObjectSubset} from 'expect/build/utils';
+import {getObjectSubset} from '@jest/expect-utils';
 import {
   DIFF_DELETE,
   DIFF_EQUAL,
